Build containerd/continuity on multiple Unix OSes
While the .go files have multiple build tags, they haven't been tested on
GitHub Actions except for FreeBSD.

This change also fixes #183.
