+++ devices/mknod_unix.go
@@ -1,5 +1,5 @@
-//go:build linux || darwin || solaris
-// +build linux darwin solaris
+//go:build !(freebsd || windows)
+// +build !freebsd,!windows
 
 /*
    Copyright The containerd Authors.
