+++ driver/lchmod_unix.go
@@ -1,5 +1,5 @@
-//go:build darwin || freebsd || solaris
-// +build darwin freebsd solaris
+//go:build darwin || freebsd || netbsd || openbsd || solaris
+// +build darwin freebsd netbsd openbsd solaris
 
 /*
    Copyright The containerd Authors.
